Aliasing ! causes subtle problems.
A test statement like the following:
  https://github.com/direnv/direnv/blob/cb5222442cb9804b1574954999f6073cc636eff0/internal/cmd/shell_bash.go#L18
caused the exclamation mark to be interpreted as
an alias for sudo instead of as a logical negation.
This may cause commands to be accidentally run as
root, expecially when sudo credentials have been
cached.
To avoid this, we simply remove the alias (even
though it is probably also better to move the
exclamation mark inside the [[ in the direnv hook).