Run ssh-to-age as a systemd user service.
Before this change we were running this from the
home-manager activation script, but on NixOS this
script runs in the environment of the systemd system
instance which means that variables like
XDG_RUNTIME_DIR are not available.
It is also not needed to have this secret available
when the user is not logged in.